CREATING AN MQ2 GAS SENSOR WITH ARDUINO UNO AND TRANSMITTING DATA TO A MOBILE PHONE VIA BLUETOOTH MODULE
The project aims to create an MQ2 Gas Sensor using Arduino Uno. The system, when activated, will have a green LED lit in the absence of gas, while in the presence of gas, a red LED will light up along with an alarm from the buzzer. Additionally, integrated into the system is an HC-06 Bluetooth module, which will send alerts to my mobile application indicating either 'GAS DETECTED' or 'NO GAS' based on the gas situation. Once the sensor is completed, it will be tested and activated.